+++ rat-salsa/src/poll/rendered.rs
@@ -1,14 +1,13 @@
 use crate::rendered::RenderedEvent;
-use crate::{AppContext, AppState, Control, PollEvents};
+use crate::{Control, PollEvents};
 use std::any::Any;
 
 /// Sends an event after a render of the UI.
 #[derive(Debug, Default)]
 pub struct PollRendered;
 
-impl<Global, State, Event, Error> PollEvents<Global, State, Event, Error> for PollRendered
+impl<Event, Error> PollEvents<Event, Error> for PollRendered
 where
-    State: AppState<Global, Event, Error> + ?Sized,
     Event: 'static + Send + From<RenderedEvent>,
     Error: 'static + Send + From<std::io::Error>,
 {
@@ -16,16 +15,12 @@ where
         self
     }
 
-    fn poll(&mut self, _ctx: &mut AppContext<'_, Global, Event, Error>) -> Result<bool, Error> {
+    fn poll(&mut self) -> Result<bool, Error> {
         // doesn't poll. it's triggered by a repaint.
         Ok(false)
     }
 
-    fn read_exec(
-        &mut self,
-        state: &mut State,
-        ctx: &mut AppContext<'_, Global, Event, Error>,
-    ) -> Result<Control<Event>, Error> {
-        state.event(&RenderedEvent.into(), ctx)
+    fn read(&mut self) -> Result<Control<Event>, Error> {
+        Ok(Control::Event(RenderedEvent.into()))
     }
 }